+++ src/diagnostics.ts
@@ -1,8 +1,41 @@
+import * as tsTypes from "typescript";
 import { red, white, yellow } from "colors/safe";
 
 import { tsModule } from "./tsproxy";
 import { RollupContext } from "./context";
-import { IDiagnostics } from "./tscache";
+import { formatHost } from "./diagnostics-format-host";
+
+export interface IDiagnostics
+{
+	flatMessage: string;
+	formatted: string;
+	fileLine?: string;
+	category: tsTypes.DiagnosticCategory;
+	code: number;
+	type: string;
+}
+
+export function convertDiagnostic(type: string, data: tsTypes.Diagnostic[]): IDiagnostics[]
+{
+	return data.map((diagnostic) =>
+	{
+		const entry: IDiagnostics = {
+			flatMessage: tsModule.flattenDiagnosticMessageText(diagnostic.messageText, "\n"),
+			formatted: tsModule.formatDiagnosticsWithColorAndContext(data, formatHost),
+			category: diagnostic.category,
+			code: diagnostic.code,
+			type,
+		};
+
+		if (diagnostic.file && diagnostic.start !== undefined)
+		{
+			const { line, character } = diagnostic.file.getLineAndCharacterOfPosition(diagnostic.start);
+			entry.fileLine = `${diagnostic.file.fileName}(${line + 1},${character + 1})`;
+		}
+
+		return entry;
+	});
+}
 
 export function printDiagnostics(context: RollupContext, diagnostics: IDiagnostics[], pretty = true): void
 {
